// src/app/services/api/model/loginApp.ts
/**
 * ApiEventos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginApp { 
    username?: string | null;
    password?: string | null;
    email?: string | null;
}